image: Remove obsolete ova file
Some methods from ova.go were moved into in the builder.
